+++ templates/parts/generator/generator-xml-footer.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Import XML file content: Footer
+ *
+ * @package    Invoices
+ * @copyright  WebMan Design, Oliver Juhas
+ *
+ * @since    1.2.0
+ * @version  1.2.0
+ */
+
+
+
+
+
+?>
+
+</channel>
+</rss>