+++ templates/parts/generator/generator-xml-header.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * Import XML file content: Header
+ *
+ * @package    Invoices
+ * @copyright  WebMan Design, Oliver Juhas
+ *
+ * @since    1.2.0
+ * @version  1.2.0
+ */
+
+
+
+
+
+?>
+
+<?xml version="1.0" encoding="UTF-8" ?>
+<rss version="2.0"
+	xmlns:excerpt="http://wordpress.org/export/1.2/excerpt/"
+	xmlns:content="http://purl.org/rss/1.0/modules/content/"
+	xmlns:wfw="http://wellformedweb.org/CommentAPI/"
+	xmlns:dc="http://purl.org/dc/elements/1.1/"
+	xmlns:wp="http://wordpress.org/export/1.2/"
+>
+<channel>
+	<wp:wxr_version>1.2</wp:wxr_version>